With these modifications Eutest.ex now will show how a test program failed and even put the contents of the error file if one exists.  Failed translations are also included in the produced html file.  A summary of failed translations is now put at the bottom.
We cannot use ifdef all of the time here because one could run eutest.ex from the windows interpreter to drive the dos interpreter executions of the various t_*.e files.  So, we check the translator filename to decide whether or not to use the -CON option.  A cc parameter has been added to allow you to specify which compiler you want the translator to use.  This will be useful if you have more than one compatible compiler installed on your system.

Eutest calls system_exec using cmd.exe instead of running a batch directly from system_exec. (Batch files with system_exec is not supported)

Problems with compiling, translating or crashing are logged in ctc.log.  Then they are integrated into the output when the log is processed.

For counter tests, control.err files are used for verifying that the interpreter is generating the ex.err correctly.