Who wants to run your software from a huge thing that must sit on your desk or be carried in your handbag?  Sure, some computers are smaller than a regularly sized notebook but some of us would like to eventually see EUPHORIA run on a phone or a wrist watch.  On the flip side, even as late as last year some big computers are using big-endian processors.  It's not dead after all...
ENTER the 'arch' option for the translator and some defines that will be defined according to what processor you are running on.  On current machines the defines will be INTEL32 and LITTLE_ENDIAN.  On a big endian machine BIG_ENDIAN will be defined and LITTLE_ENDIAN will not be defined.

The idea is that we should make the translator ready to be able to create the C source for any new supported architecture using an option from any machine that supports translation to C.

The EUPHORIA interpreter or translator should set defines, they are symbols used with ifdef-syntax, for EUPHORIA programs to have different behavior that is determined by various processor characteristics.  Such as, Byte-sex, Intel 386 compatibility, and word size.

RULES:
In this branch all sources should be translatable by the main-trunk version as well as itself.  Yet the demos and unit tests will assume the tool kit compiled from the last revision in this branch.

In this revision:
* Updated documentation to mention INTEL32 and LITTLE_ENDIAN.
* In both the Interpreter and Translator INTEL32 and LITTLE_ENDIAN are always defined.
* Got rid of constants UWIN32, ULINUX, etc...  They are redundant since FREEBSD != LINUX now.
* Fixed a problem in std/filesys.e where NETBSD will error out.
* Made std/io.e endian neutral (except we need to put with define LITTLE_ENDIAN for compatibility with the trunk)
* Wrote  get_machine_architecture_defines for returning the defines we want.
* Created machine_func(M_DEFINES,...) for getting the defines
* Added -arch option to the translator
* Fixed a problem with the translator trying to parse its options twice.  (which clobbered flags)
* Added a unit test t_arch.e